requirements.txt
python-telegram-bot
Flask
sentence-transformers
requests
pdfplumber
python-docx
